+++ xims/lib/XIMS/Exporter/VLibraryItem/Document.pm
@@ -88,7 +88,8 @@ sub remove {
         grantee_id => XIMS::PUBLICUSERID(),
         content_id => $self->{Object}->id()
     );
-    $privs_object->delete();
+    
+    $privs_object->delete() if $privs_object;
 
     # unpublish published children
     my @children = $self->{Object}->children_granted(
